scorm-export: load Node-side assets from one place
build.mjs and the test helper each had their own copy of "read the templates,
the vendored bridge and lz-string off disk" — the helper's comment literally
said it did it "the same way build.mjs does", which is the sort of duplication
that drifts. Both now call loadNodeAssets().

It lives in src/node-assets.js rather than src/index.js because the browser
bundles index.js, which therefore has to stay free of fs. It uses createRequire
rather than import.meta.resolve so it works both under plain Node and under
Vite's SSR transform, which the tests run through.

Output is unchanged: a package built after this refactor is byte-identical to
one built before it.
